Film cards fixes (1st batch) (#24)
Added handling of the media object being a film to the carousel. It seems to detect them properly, but what is needed is proper loading of the metadata. What needs more work is parsing and combining data from the junction tables that map many-to-many relationships, not only for films, but for other data too.

In case of films, these are:
1. `film_posters` – though I'm not sure this table is needed. The static files handler package (`cdn`, name chosen for brevity) could possibly create directories with all the posters per file, with some key-value store of the average votes for each poster. This could be faster and simpler than using DB for this.
2. `film_countries`
3. `cast`, `countries`, `episodes`, `langs`, `seasons` tables for TV
4. `*cast` and `studio*` tables in the `people` schema

This is the first batch of the fixes. It seems that the random media carousel does not handle multiple media types too well, because we may get something like this:
```
mediaItem is a film MediaCarousel.svelte:269:13
mediaItem:
Object { UUID: undefined, kind: "album"
```

Reviewed-on: https://codeberg.org/mjh/LibRate/pulls/24